lineup special

+++ console/migrations/m190113_081758_team_country_news.php
@@ -15,7 +15,7 @@ class m190113_081758_team_country_news extends Migration
      */
     public function safeUp()
     {
-        $this->addColumn(self::TABLE, 'team_news_id', $this->integer()->defaultValue(0));
+        $this->addColumn(self::TABLE, 'team_news_id', $this->integer(11)->defaultValue(0));
         $this->dropColumn(self::TABLE_2, 'user_country_news_id');
     }
 
@@ -24,7 +24,7 @@ class m190113_081758_team_country_news extends Migration
      */
     public function safeDown()
     {
-        $this->addColumn(self::TABLE_2, 'user_country_news_id', $this->integer()->defaultValue(0));
+        $this->addColumn(self::TABLE_2, 'user_country_news_id', $this->integer(11)->defaultValue(0));
         $this->dropColumn(self::TABLE, 'team_news_id');
     }
 }
